Support building without shared disk support
